+++ CHANGELOG.md
@@ -4,14 +4,24 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 ## [unreleased]
 
+### Removed
+
+- `sandstone/WizardPanels` prop `footer`
+
+### Changed
+
+- `sandstone/WizardPanels` prop `buttons` to be `footer`
+
 ### Added
 
-- `sandstone/ImageItem` public class names `horizontal` and `vertical`
 - `sandstone/DatePicker` function `dateToLocaleString` to create locale-aware date strings
+- `sandstone/ImageItem` public class names `horizontal` and `vertical`
+- `sandstone/PopupTabLayout` and `sandstone/TabLayout` prop `onTabAnimationEnd` to notify consumers when the animation to collapse or expand the tabs completes
 - `sandstone/TimePicker` function `timeToLocaleString` to create locale-aware time strings
 
 ### Fixed
 
+- `sandstone/Scroller` and `sandstone/VirtualList` to handle a keydown event without an error
 - `sandstone/Panels` to not fire transition events when initially rendered
 - `sandstone/Tooltip` style to match latest designs
 - `sandstone/VirtualList` to support navigation with spottable children inside an item